Rename replicate_image with process_job.
Rename job.py to replication_job.py.
