+++ mash/services/replication/service.py
@@ -83,7 +83,7 @@ class ReplicationService(PipelineService):
         Replicate image for job that matches job_id.
         """
         job = self.jobs[job_id]
-        job.replicate_image()
+        job.process_job()
 
     def _get_listener_msg_args(self):
         """
